scrapers/company_scrapers.py
"""
Company Career Page Scrapers

Each major company has a different career page structure.
This module contains scraper functions for priority companies.
"""

import requests
from bs4 import BeautifulSoup
import time
import hashlib
import sys
import os
import json

sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from utils.database import job_exists, save_job
from utils.scorer import score_job

def scrape_wells_fargo():
    """Scrape Wells Fargo careers page."""
    
    # Wells Fargo uses Workday
    # Their API endpoint for job search
    api_url = "https://wd1.myworkdaysite.com/recruiting/wf/WellsFargoJobs"
    
    new_jobs = []
    
    try:
        print("Scraping Wells Fargo careers...")
        
        # Wells Fargo search parameters
        # This is a simplified example - may need adjustment based on their actual API
        search_url = f"{api_url}?q=credit%20analyst&locations=0c8a2e026ceb01d9e42f1e96890f0000"
        
        headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
        
        response = requests.get(search_url, headers=headers, timeout=10)
        
        # Note: Wells Fargo uses dynamic loading
        # Full implementation would require Selenium or their API endpoint
        print("  ℹ️  Wells Fargo requires advanced scraping (Workday platform)")
        print("  Recommendation: Use LinkedIn/Indeed filters for Wells Fargo instead")
        
    except Exception as e:
        print(f"  ❌ Error: {str(e)}")
    
    return new_jobs

def scrape_jpmorgan():
    """Scrape JPMorgan careers page."""
    
    # JPMorgan uses iCIMS
    base_url = "https://jpmc.fa.oraclecloud.com/hcmUI/CandidateExperience/en/sites/CX_1001"
    
    new_jobs = []
    
    try:
        print("Scraping JPMorgan careers...")
        
        # JPMorgan has a jobs API
        # Example search: corporate banking analyst
        search_params = {
            'keyword': 'credit analyst',
            'location': 'United States'
        }
        
        # Note: JPMorgan requires complex API calls with authentication tokens
        print("  ℹ️  JPMorgan requires advanced scraping (Oracle platform)")
        print("  Recommendation: Use LinkedIn/Indeed filters for JPMorgan instead")
        
    except Exception as e:
        print(f"  ❌ Error: {str(e)}")
    
    return new_jobs

def scrape_keybank():
    """Scrape KeyBank careers page."""
    
    # KeyBank uses Workday
    base_url = "https://keybank.wd5.myworkdayjobs.com/External_Career_Site"
    
    new_jobs = []
    
    try:
        print("Scraping KeyBank careers...")
        
        # Workday platforms require specific API calls
        print("  ℹ️  KeyBank uses Workday - requires advanced scraping")
        print("  Recommendation: Check their careers page manually or use aggregators")
        
    except Exception as e:
        print(f"  ❌ Error: {str(e)}")
    
    return new_jobs

def scrape_bmo():
    """Scrape BMO careers page."""
    
    base_url = "https://jobs.bmo.com/us/en"
    
    new_jobs = []
    
    try:
        print("Scraping BMO careers...")
        
        # BMO has a public job search
        search_url = f"{base_url}/search-results?keywords=credit%20analyst"
        
        headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
        
        response = requests.get(search_url, headers=headers, timeout=10)
        soup = BeautifulSoup(response.content, 'html.parser')
        
        # Note: Actual parsing would depend on BMO's page structure
        print("  ℹ️  BMO page structure needs manual inspection")
        
    except Exception as e:
        print(f"  ❌ Error: {str(e)}")
    
    return new_jobs

def scrape_generic_workday(company_name, workday_url, keywords):
    """
    Generic scraper for Workday-based career sites.
    
    Args:
        company_name: Name of the company
        workday_url: Base Workday URL (e.g., https://company.wd1.myworkdayjobs.com/CareerSite)
        keywords: List of search keywords
    """
    
    new_jobs = []
    
    try:
        print(f"Scraping {company_name} (Workday)...")
        
        # Workday sites typically have this structure
        # But require GraphQL API calls with specific tokens
        # This is a placeholder for the full implementation
        
        print(f"  ℹ️  {company_name} uses Workday platform")
        print("  Note: Workday scrapers require Selenium + API reverse engineering")
        
    except Exception as e:
        print(f"  ❌ Error: {str(e)}")
    
    return new_jobs

# Company configuration
COMPANIES = {
    'Wells Fargo': {
        'url': 'https://wd1.myworkdaysite.com/recruiting/wf/WellsFargoJobs',
        'platform': 'Workday',
        'scraper': scrape_wells_fargo
    },
    'JPMorgan': {
        'url': 'https://jpmc.fa.oraclecloud.com/hcmUI/CandidateExperience/en/sites/CX_1001',
        'platform': 'Oracle',
        'scraper': scrape_jpmorgan
    },
    'KeyBank': {
        'url': 'https://keybank.wd5.myworkdayjobs.com/External_Career_Site',
        'platform': 'Workday',
        'scraper': scrape_keybank
    },
    'BMO': {
        'url': 'https://jobs.bmo.com/us/en',
        'platform': 'Custom',
        'scraper': scrape_bmo
    }
}

def scrape_all_companies():
    """Run scrapers for all configured companies."""
    
    all_new_jobs = []
    
    for company, config in COMPANIES.items():
        jobs = config['scraper']()
        all_new_jobs.extend(jobs)
        time.sleep(5)
    
    return all_new_jobs

if __name__ == '__main__':
    from utils.database import init_database
    
    init_database()
    
    print("\n" + "="*60)
    print("COMPANY CAREER PAGE SCRAPERS")
    print("="*60 + "\n")
    
    new_jobs = scrape_all_companies()
    
    print("\n" + "="*60)
    print(f"SUMMARY: Found {len(new_jobs)} new jobs")
    print("="*60)
    print("\nNOTE: Most major banks use complex platforms (Workday, Oracle)")
    print("For best results, use LinkedIn/Indeed scrapers with company filters")
